#AB37 - Rent Admin change password

+++ clover/src/app/app.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { User } from './entities/User/user';
+import { UserDetailsService } from './services/userDetails/user-details.service';
 
 @Component({
   selector: 'app-root',
@@ -9,12 +11,14 @@ import { Router } from '@angular/router';
 export class AppComponent {
   title = 'clover';
   emaill:string;
+  user: User;
+  userId: number;
 
   ngOnInit(): void {
     
   }
 
-  constructor(public router: Router){
+  constructor(public router: Router, public userService: UserDetailsService){
     
   }
   
@@ -23,7 +27,21 @@ export class AppComponent {
     return !!localStorage.getItem('regId');
   }
   LogOut(){
-    localStorage.clear();
-    this.router.navigateByUrl("/");
+    this.userId = JSON.parse(localStorage.getItem("regId"));
+    this.userService.getUserById(this.userId).toPromise().then(
+      dataV => {
+        this.user = dataV;
+        this.userService.refreshList();
+      }
+    )
+
+
+    if(this.user.logOut == false){
+      alert("Morate prvo promeniti lozinku");
+    }
+    else{
+      localStorage.clear();
+      this.router.navigateByUrl("/");
+    }
   }
 }
